fix 64bit
git-svn-id: https://svn.php.net/repository/pecl/intl/trunk@256785 c90b9560-bf6c-de11-be94-00142212c4b1

+++ dateformat/dateformat_attr.c
@@ -296,7 +296,7 @@ PHP_FUNCTION( datefmt_set_pattern )
 PHP_FUNCTION( datefmt_get_locale )
 {
 	char *loc;
-	int  loc_type =ULOC_ACTUAL_LOCALE;
+	long  loc_type =ULOC_ACTUAL_LOCALE;
 
 	DATE_FORMAT_METHOD_INIT_VARS;
